加入：乘除法混合运算、四则混合运算2、小数加减运算

+++ bsd-shuxue-3/Domain/Framework.cs
@@ -22,7 +22,7 @@ namespace bsd_shuxue_3.Domain
         /// <summary>
         /// 问题答案
         /// </summary>
-        int Answer { get; }
+        Decimal Answer { get; }
     }
 
     /// <summary>
